feat(019): add running costs to budget period view
Phase 3 (US4): Add getRunningCostsForPeriod() helper using authoritative
anthropic_workspace_costs data. Budget period detail now shows running
costs section with last-updated timestamp, visually distinct from billed
costs. Completes T012-T014.
